@@ -5,40 +5,34 @@
   "packageManager": "pnpm@10.17.0",
   "dependencies": {
     "@biomejs/biome": "^2.2.4",
-    "@sentry/browser": "^5.15.5",
-    "action-reducer": "^0.3.1",
-    "axios": "^0.19.2",
-    "bulma": "^0.8.2",
-    "connected-react-router": "^6.8.0",
-    "history": "^4.10.1",
-    "immer": "^6.0.5",
-    "raven-for-redux": "^1.4.0",
-    "raven-js": "^3.27.2",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-ga": "^2.7.0",
-    "react-redux": "^7.2.0",
-    "react-router": "^5.1.2",
-    "react-router-dom": "^5.1.2",
-    "react-twitter-widgets": "^1.7.1",
-    "redux": "^4.0.5",
+    "@sentry/browser": "^10.12.0",
+    "@sentry/react": "^10.12.0",
+    "action-reducer": "^0.4.0",
+    "axios": "^1.12.2",
+    "bulma": "^1.0.4",
+    "immer": "^10.1.3",
+    "react": "^19.1.1",
+    "react-dom": "^19.1.1",
+    "react-ga": "^3.3.1",
+    "react-redux": "^9.2.0",
+    "react-router-dom": "^7.9.1",
+    "react-twitter-widgets": "^1.11.0",
+    "redux": "^5.0.1",
     "redux-logger": "^3.0.6",
-    "redux-saga": "^1.1.3"
+    "redux-saga": "^1.3.0"
   },
   "devDependencies": {
-    "@types/jest": "^25.2.1",
-    "@types/node": "^13.13.5",
-    "@types/raven-for-redux": "^1.1.1",
-    "@types/react": "^16.9.34",
-    "@types/react-dom": "^16.9.7",
-    "@types/react-redux": "^7.1.8",
-    "@types/react-router-dom": "^5.1.5",
-    "@types/redux-logger": "^3.0.7",
-    "@vitejs/plugin-react": "^5.0.0",
+    "@types/jest": "^30.0.0",
+    "@types/node": "^24.5.2",
+    "@types/react": "^19.1.13",
+    "@types/react-dom": "^19.1.9",
+    "@types/react-redux": "^7.1.34",
+    "@types/redux-logger": "^3.0.13",
+    "@vitejs/plugin-react": "^5.0.3",
     "jsdom": "^27.0.0",
     "typescript": "5.9.2",
-    "vite": "^5.0.0",
-    "vitest": "^1.0.0"
+    "vite": "^7.1.6",
+    "vitest": "^3.2.4"
   },
   "scripts": {
     "start": "vite --config vite.config.mts",
