@@ -1,9 +1,7 @@
-import React from 'react';
-import { render, unmountComponentAtNode } from 'react-dom';
 import Card from './Card';
+import renderWithRoot from '../test/render';
 
 it('renders without crashing', () => {
-  const div = document.createElement('div');
-  render(<Card title="test" body={['test', 'test2']} link="https://8398a7.github.io" />, div);
-  unmountComponentAtNode(div);
+  const { unmount } = renderWithRoot(<Card title="test" body={['test', 'test2']} link="https://8398a7.github.io" />);
+  unmount();
 });
