@@ -1,15 +1,12 @@
-import { connectRouter } from 'connected-react-router';
-import type { History } from 'history';
 import { combineReducers } from 'redux';
 import { fork } from 'redux-saga/effects';
 import meta, { metaSaga } from './Meta';
 import ui from './UI';
 
-const rootReducer = (history: History) =>
+const rootReducer = () =>
   combineReducers({
     meta,
     ui,
-    router: connectRouter(history),
   });
 export default rootReducer;
 export type RootState = ReturnType<ReturnType<typeof rootReducer>>;
