@@ -1,13 +1,8 @@
-import { routerMiddleware } from 'connected-react-router';
-import type { History } from 'history';
-import createRavenMiddleware from 'raven-for-redux';
-import Raven from 'raven-js';
-import { applyMiddleware, compose, createStore } from 'redux';
+import { applyMiddleware, compose, legacy_createStore as createStore } from 'redux';
 import type { Middleware, StoreEnhancer } from 'redux';
 import { createLogger } from 'redux-logger';
 import createSagaMiddleware from 'redux-saga';
 import createRootReducer, { rootSaga } from './ducks';
-import { googleAnalytics } from './lib/reactGAMiddleware';
 
 type ReduxDevToolsExtension = () => StoreEnhancer;
 
@@ -18,29 +13,21 @@ type ExtendedWindow = Window &
 
 const identityEnhancer: StoreEnhancer = (next) => next;
 
-export default (history: History, dsn: string) => {
+const configureStore = () => {
   const sagaMiddleware = createSagaMiddleware();
-  const middlewares: Middleware[] = [];
+  const middlewares: Middleware[] = [sagaMiddleware];
 
   if (import.meta.env.MODE !== 'production') {
     middlewares.push(createLogger());
   }
 
-  Raven.config(dsn).install();
-
   const devtoolsExtension = (window as ExtendedWindow).__REDUX_DEVTOOLS_EXTENSION__;
   const devtools = import.meta.env.MODE !== 'production' && devtoolsExtension ? devtoolsExtension() : identityEnhancer;
 
   const store = createStore(
-    createRootReducer(history),
+    createRootReducer(),
     compose(
-      applyMiddleware(
-        routerMiddleware(history),
-        googleAnalytics,
-        sagaMiddleware,
-        createRavenMiddleware(Raven, {}),
-        ...middlewares,
-      ),
+      applyMiddleware(...middlewares),
       devtools,
     ),
   );
@@ -49,3 +36,5 @@ export default (history: History, dsn: string) => {
 
   return store;
 };
+
+export default configureStore;
